symbol link for change config

+++ src/UChainApp/ucd/executor.cpp
@@ -333,17 +333,10 @@ void executor::initialize_output()
 
     auto file = metadata_.configured.file.string().compare("uc.conf") \
                         ? metadata_.configured.file:default_data_path() / metadata_.configured.file;
-    if(metadata_.configured.file.string().compare("uc.conf"))
+    if(!metadata_.configured.file.string().empty() && metadata_.configured.file.string().compare("uc.conf"))
     {
-        if(boost::filesystem::exists(default_data_path()/"conf"))
-        {
-            boost::filesystem::remove(default_data_path()/"conf");
-        }
-    #ifdef __WIN32__
-        CreateSymbolicLinkA(default_data_path()/"conf", metadata_.configured.file.string(), 0) ;
-    #else
-        symlink(metadata_.configured.file.string().c_str(), (default_data_path()/"conf").string().c_str());
-    #endif
+        boost::filesystem::remove(default_data_path()/"conf");  
+        boost::filesystem::create_symlink(metadata_.configured.file, default_data_path()/"conf");
     }
    
     if (file.empty())
